no RAM size check with esos_persist option
Disable the RAM size check if esos_persist kernel commandline option is used.
Enforcing this ram amount is relevant only if ESOS needs to run from RAM.
If ESOS is run with esos_persist kernel parameter, it is not doing that.

With this change, people that want to use ESOS with less than 3GB of RAM
only need to edit the grub config in their boot partition to add esos_persist
and they are set, no need to recompile/rebuild initramfs.